added logging and more
